pyproject.toml
[project]
name = "ria-voice-assistant"
version = "1.0.0"
description = "BlueStone voice AI jewelry consultant powered by ElevenLabs and Twilio"
requires-python = ">=3.11"
dependencies = [
    "fastapi==0.115.12",
    "uvicorn[standard]==0.34.2",
    "redis==5.2.1",
    "twilio==9.6.3",
    "httpx==0.28.1",
    "pydantic==2.11.4",
    "pydantic-settings==2.9.1",
    "python-dotenv==1.1.0",
    "elevenlabs==1.54.0",
    "python-multipart>=0.0.27",
    # cross-call memory: customers + conversations in Postgres, Gemini for the per-call summariser
    "sqlalchemy>=2.0.30",
    "psycopg2-binary>=2.9.9",
    "google-genai>=0.8.0",
    "alembic>=1.13.0",
]

[project.optional-dependencies]
eval = [
    "streamlit>=1.40.0",
    "pandas>=2.2.0",
]

[project.scripts]
ria = "main:run"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
include = ["app", "eval", "migrations", "main.py", "alembic.ini"]
